Fix migration bug in SVector
When inserting into a sorted SVector we check if it's necessary
to migrate the SVector to use an RB tree, and the code might
proceed to do so. However, the length of the SVector is
incremented before the migration, which causes the migrator to
read one index past the current last item. This is practically
zeroed in almost every possible case but it's still wrong.