Fix inserting node_id multiple times as an adj
`Svector` should only increment the size count if something was
inserted and refuse to insert the same data twice into an ordered
vector.

Currently if for example a children node_id list on modify
contains the same node_id more than once it makes `SVector_Size()`
and the actual vector size differ. This bug in turn breaks many
things already but it will also eventually completely break the
RDB serialization and deserialization and cause a crash.