Optimize substring lookup
Turn `strnstr()` into a new function called `strnstrn()` that
avoids finding the length of the find string because it seems
to be a rather hot part of the code.

`strnstrn()` seems to benefit from being marked as a hot
function, at least for now. Eventually instruction cache
optimizations will break because all the hot functions
can't fit the cache lines nor be close enough to each
other when needed. Perhaps in the future we should determine
which hot functions are commonly used together and create
more specific hot sections in the elf.