Add Trx_Sync() to keep last visited stats for free
The database optimizations like indexing and subtree compression
would greatly benefit from being able to determine if nodes and
subtrees are visited often. We can utilize transaction ids for
that purpose, as we can determine the distance of the
current/next to the label id, which is somewhat proportional to
time.

Currently we only use transactions for tracking whether nodes
should be visited within a traversal or in a nested traversal.
The transaction colors are a scarce resource and not needed if we
just want to look a the ids. Therefore, we need to a new function
for just syncing the current transaction id to nodes when we don't
actually traverse.