Retrieve nested references in find
Previously we allowed retrieving fields from a single-ref field
but it's often useful to allow retrieving all fields from an edge
field point to multiple nodes, e.g. dereferencing a `references`
field. The client must be careful to not retrieve a too big field
as `limit` does not apply when dereferencing edge fields.