Redis alloc functions can never fail
Redis will panic on a fatal OOM and it can never return a NULL
pointer. This should never happen in Redis because it should be
able to signal us about the upcoming OOM, given that we'd be
listening to such signals. If we are ever to replace the memory
allocation functions, it would be wise to follow similar
principles as it simplifies things a lot.