Fix != filter operator when $language is set
If `$language` is set we always assume that any field containg
a string value might be a `text` field. This is problematic with
the `!=` operator in a filter because `field.lang` doesn't exist
for `string` fields.

For example:

```
$filter: {
  $field: 'type',
  $operator: '!=',
  $value: 'league',
}
```

currently becomes:

```
[ '$2 $1 f c L $4 $3 f c L N', 'type', 'league', 'type.en', 'league' ]
```

while it should either skip `.en` for non-text fields or change the
OR operator `N` to the AND operator `M` when `$operator` in the
`$filter` equals to `!=`.