SO: Better handling of last part array index
Most SelvaOjbect set functions should handle the case when the last
part of the field name is an index. Indicies in the middle of the
field name are already handled by `get_key` functions but the last
part is tricky because `get_key()` will return the key of the
array.

It would be also possible to optimize this further in the future by
doing more of the work in `get_key()` and signaling the results in a
better way to the caller.

https://github.com/atelier-saulx/based-core/issues/152